requirements.txt
mlflow
optuna
lightgbm
scikit-learn
pandas
pyyaml
fastapi
uvicorn
pydantic
